refactor(proactive): extract state and capability domains
